fix(shipping): CHECKOUT-8872 Move Helper Method
