+++ packages/core/src/app/shipping/utils/index.ts
@@ -1,1 +1,2 @@
 export { generateItemHash } from './generateItemHash';
+export { setRecommendedOrMissingShippingOption } from './setRecommendedOrMissingShippingOption';
